+++ src/Installer.php
@@ -18,46 +18,70 @@ class Installer extends LibraryInstaller
     
     const LUYA_FILE = 'luyadev/installer.php';
     
-    private $_relativeVendorDir;
-    
+    /**
+     * {@inheritDoc}
+     */
     public function supports($packageType)
     {
         return $packageType == 'luya-core' || $packageType == 'luya-extension' || $packageType == 'luya-module';
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function install(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
         // install the package the normal composer way
         parent::install($repo, $package);
-        
         $this->addPackage($package);
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function update(InstalledRepositoryInterface $repo, PackageInterface $initial, PackageInterface $target)
     {
         parent::update($repo, $initial, $target);
-        
         $this->removePackage($initial);
         $this->addPackage($target);
     }
     
+    /**
+     * {@inheritDoc}
+     */
     public function uninstall(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
         parent::uninstall($repo, $package);
-        
         $this->removePackage($package);
     }
     
+    /**
+     * Add a package to installer
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function addPackage(PackageInterface $package)
     {
         $this->writeInstaller($this->addConfig($package));
     }
     
+    /**
+     * Remove a package from installer
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function removePackage(PackageInterface $package)
     {
         $this->writeInstaller($this->removeConfig($package));
     }
     
+    /**
+     * Get the installer array
+     *
+     * @return array
+     */
     protected function getInstallers()
     {
         $file = $this->vendorDir . DIRECTORY_SEPARATOR . self::LUYA_FILE;
@@ -75,6 +99,13 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Ensure a config for a package.
+     *
+     * @param PackageInterface $package
+     * @param array $config
+     * @return void
+     */
     protected function ensureConfig(PackageInterface $package, array $config)
     {
         $packageConfig = [
@@ -91,7 +122,13 @@ class Installer extends LibraryInstaller
         
         return $packageConfig;
     }
-        
+     
+    /**
+     * Remove a package from the config
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function removeConfig(PackageInterface $package)
     {
         $data = $this->getInstallers();
@@ -103,6 +140,12 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Get the LUYA extra binary data.
+     *
+     * @param PackageInterface $package
+     * @return array
+     */
     protected function getPackageExtraData(PackageInterface $package)
     {
         if (empty($package->getExtra())) {
@@ -112,6 +155,12 @@ class Installer extends LibraryInstaller
         return isset($package->getExtra()[self::LUYA_EXTRA]) ? $package->getExtra()[self::LUYA_EXTRA] : [];
     }
     
+    /**
+     * Add a package to the config
+     *
+     * @param PackageInterface $package
+     * @return void
+     */
     protected function addConfig(PackageInterface $package)
     {
         $data = $this->getInstallers();
@@ -120,6 +169,12 @@ class Installer extends LibraryInstaller
         return $data;
     }
     
+    /**
+     * Write the installer.php file in vendor folder
+     *
+     * @param array $data
+     * @return void
+     */
     protected function writeInstaller(array $data)
     {
         $file = $this->vendorDir . DIRECTORY_SEPARATOR . self::LUYA_FILE;
@@ -138,6 +193,8 @@ class Installer extends LibraryInstaller
             @opcache_invalidate($file, true);
         }
     }
+
+    private $_relativeVendorDir;
     
     /**
      * Read the relative vendor-dir from composer config.
@@ -153,6 +210,4 @@ class Installer extends LibraryInstaller
         
         return $this->_relativeVendorDir;
     }
-    
-    
 }