+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file. This project make usage of the [Yii Versioning Strategy](https://github.com/yiisoft/yii2/blob/master/docs/internals/versions.md).
 
-## 1.0.4 (in progress)
+## 1.0.4 (12. December 2018)
 
 + [#7](https://github.com/luyadev/luya-composer/issues/7) Add option `symlink` in luya extra section to disable the symlinking of the luya binary into the application folder.
 + [#2](https://github.com/luyadev/luya-composer/issues/2) Relative vendor directory and alias for block paths.